% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grib2ANetCDF.R
\name{Grib2ANetCDF}
\alias{Grib2ANetCDF}
\title{Transforma archivos .grib2 a .nc}
\usage{
Grib2ANetCDF(
  ruta_in,
  ruta_out,
  parallel = FALSE,
  ncores = detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{ruta_in}{Ruta del archivo de entrada \code{.grib2} o un vector de rutas si se ejecuta en paralelo.}

\item{ruta_out}{Ruta del archivo de salida \code{.nc} o un vector de rutas de salida si se ejecuta en paralelo.}

\item{parallel}{Lógico que indica si se debe ejecutar en paralelo (TRUE) o secuencialmente (FALSE).}

\item{ncores}{Número de núcleos a utilizar para la ejecución paralela, solo relevante si parallel = TRUE.}
}
\value{
No devuelve un valor explícitamente, pero ejecuta un comando en el
sistema que transforma el archivo y lo guarda en la ubicación especificada.
}
\description{
Transforma archivos .grib2 a .nc
}
\details{
Esta función requiere que el Subsistema de Windows para Linux (WSL) esté instalado
y configurado correctamente en el sistema, y que el script de shell \code{transform_0125.sh}
esté disponible en la ruta especificada dentro de WSL.
}
\examples{
Grib2ANetCDF("C:/path/to/input/file.grib2", "C:/path/to/output/file.nc")
}
