% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesarDirectorios.R
\name{procesarDirectorios}
\alias{procesarDirectorios}
\title{Procesar Directorios para Transformación de Datos Meteorológicos}
\usage{
procesarDirectorios(directorioBase)
}
\arguments{
\item{directorioBase}{Ruta al directorio base que se procesará.
Se espera que este directorio contenga archivos \code{.bz2} y \code{.grib2}
para su extracción y transformación, respectivamente.}
}
\description{
Esta función automatiza el proceso de búsqueda y transformación de archivos
meteorológicos. Primero, extrae archivos \code{.bz2} encontrados en el directorio
base y sus subdirectorios. Luego, convierte archivos \code{.grib2} a formato \code{.nc}.
}
\details{
La función primero busca todos los archivos \code{.bz2} en el directorio
base y los subdirectorios, los extrae, y luego busca archivos \code{.grib2} para
convertirlos a formato \code{.nc} utilizando la función \code{Grib2ANetCDF}. Es importante
que el directorio base tenga permisos adecuados para la lectura de archivos y
escritura de los archivos transformados.

Esta función depende de las funciones \code{extraerBZ2} y \code{Grib2ANetCDF} para
realizar las tareas de extracción y transformación, respectivamente.
}
\examples{
procesarDirectorios("D:/INIA/ICON/prueba_grib2_to_nc/")

}
